+++ example/angular-5-app/src/app/other/other.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+import { Service } from '../service';
+
+@Component({
+    selector: 'app-other',
+    templateUrl: './other.component.html',
+    styleUrls: ['./other.component.css']
+})
+export class OtherComponent {
+    title = 'otherComponent';
+
+    constructor(private s: Service) {
+        this.title = s.method() ? this.title + s.method() : this.title;
+    }
+}